Ticket overview paging added
A paging system has been created for the ticket overview view. It can be customised to show any number of tickets per page, currently it is set to 100. In addition, data that is not set in a ticket is not shown in the individual ticket view, to try and make the view clearer.

--New Features
-Ticket overview paging system added.
-Data is hidden if it is not set for a ticket in the individual ticket view.
-Creation time added to table of tickets in the overview.
-Ticket overview is now coloured on alternate rows to aid in clarity.