Server-side bash script test added, positive user auth test added,  minor changes to styling
In order to isolate the front-end from testing, curl commands are used in a bash scripts to test the server-side code for authenticating users and requesting ticket data from the API. A test for a positive result of user authentication has been added to the front end testing view. A minor change was made to make the testing pages more readable.

--New Features:
-Bash script added for isolation testing the server-side user auth using curl
-Bash script added for isolation testing the server-side ticket retrieval using curl
-Positive-result test added to the front end user auth testing view.

--Changes
-Inline routing templates have been renamed for a clearer naming convention.
-Small styling change to make the testing frontend views more readable